Add all needed files and create grid
Create container div in HTML for javascript to add the grid to.
Create the divs with for loop.
Set styles in css for container to flex to use flexbox and also set wrap.
Set everything needed for divs but to make it work set the 16th+1 div
element to width 100% and border and height to 0 to force a line break