Finally add function to remove old grid
Use innerHTML to make the createGrid function to remove the old grid
when a new one is prompted by the user by overwriting the HTML.
Fix some px issues that made the grid not be the right size, now always
has same px in height and width.
Add the updateOpactiy and randomRGB function to now also make it so that
the opacity is changed by 0.1 every mouseover so the grid-item gets
darker and the random color gets more clear.
randomRGB function creates that random color every mouseover